fixing warnings. changing all shorts to ints since they are same byte size for xc8 compiler

+++ robot/stepper.c
@@ -1,14 +1,14 @@
 #include "stepper.h"
 
-unsigned short scan_360_ccw_step_count = 0;     //counter to count how many half steps since closest scanned object
-unsigned short new_adc_distance = 0;            //variable to store latest reading of adc distance
-unsigned short closest_adc_distance = 0;        //variable to store closest reading of adc distance since push button press
+unsigned int scan_360_ccw_step_count = 0;     //counter to count how many half steps since closest scanned object
+unsigned int new_adc_distance = 0;            //variable to store latest reading of adc distance
+unsigned int closest_adc_distance = 0;        //variable to store closest reading of adc distance since push button press
 
 unsigned char CW_control_byte = 0b00001111;     //stepper motor control byte for; enabled, clockwise, half-steps
 unsigned char CCW_control_byte = 0b00001101;    //stepper motor control byte for; enabled, counterclockwise, half-steps
 
 //rotate stepper CW 360 degrees. scan adc distance each half step.
-void scan360 (unsigned short steps){
+void scan360 (unsigned int steps){
     resetADC();
     spi_transfer(CW_control_byte);
     
@@ -21,7 +21,7 @@ void scan360 (unsigned short steps){
 }
 
 //move stepper CW
-void moveCW (unsigned short steps) {
+void moveCW (unsigned int steps) {
     spi_transfer(CW_control_byte);
 	for(steps; steps!=0; steps--){
         SM_STEP();
@@ -30,7 +30,7 @@ void moveCW (unsigned short steps) {
 }
 
 //move stepper CCW
-void moveCCW (unsigned short steps) {
+void moveCCW (unsigned int steps) {
     spi_transfer(CCW_control_byte);
 	for(steps; steps!=0; steps--){
         SM_STEP();
